refactor: Feature First cleanup — remove legacy cardflix code and lift shared UI to lib/shared/
- Delete all legacy app/ screens (cardflix, home, video) and their controllers
- Delete unused utils (command.dart, result.dart), provider.dart, and old model abstraction
- Remove CardflixData model and its Hive type adapter; replace with WorkoutPlanTypeAdapter
- Move Label, TopDrawer, Loading widgets to lib/shared/widgets/ with Dart 3 super-parameters
- Move color constants to lib/shared/theme/colors.dart (renamed appColors → difficultyColors)
- Move Model abstract class to lib/core/domain/model.dart
- Add restSeconds to SetConfig and warmupNote to DifficultyTier to match JSON schema
- Update all feature imports to reference lib/shared/ paths
- Fix deprecated APIs: withOpacity → withValues, network() → networkUrl()
- Upgrade pubspec to Dart 3 SDK constraint, add flutter_lints, pin video_player_android override
- Fix all flutter analyze warnings (super-parameters, unused fields, unnecessary containers)
- Update test fixture count (8 → 11) and replace boilerplate widget test with smoke test
