--- lib/app/helpers/abstraction/controller.dart
@@ -1,15 +1,0 @@
-/// External package
-import 'package:flutter/material.dart';
-
-/// Controller to the Provider widget
-abstract class Controller {
-  late final BuildContext context;
-  /// Initialize controller
-  void init() /** optional */ {}
-  /// [Provider] mounted
-  void mounted() /** optional */ {}
-  /// Build [Provider]
-  void build() /** optional */ {}
-  /// Clean data from Android/iOS
-  void dispose() /** optional */ {}
-}
