--- lib/app/helpers/color.dart
@@ -1,7 +1,0 @@
-import 'package:flutter/material.dart';
-
-const appColors = <Color>[
-  Color(0xFF30D407),
-  Color(0xFFEBE109),
-  Color(0xFFEF4040)
-];
