--- lib/app/helpers/data/cardflix.dart
@@ -1,225 +1,0 @@
-class CardflixExamples {
-  static const bumbumNaLua = {
-    "id": "4y73y475y3884y3",
-    "name": "Bumbum na lua",
-    "description": "Para voçê que busca massa muscular no bumbum.",
-    "cover":
-        "https://firebasestorage.googleapis.com/v0/b/myself-dg.appspot.com/o/powerflix%2Fbumbum_na_lua.jpg?alt=media&token=49eb1a02-4554-450a-9c9e-24357a6135ff",
-    "modules": [
-      {
-        "level": "LIGHT",
-        "description":
-            "Movimentos lentos e concentrados para otimizar o padrão de movimento e seus resultados",
-        "frequency": {"repetition": 11, "series": 3},
-        "exercises": [
-          {
-            "order": 1,
-            "name": "Agachamento com peso corporal",
-            "link":
-                "https://api-files.sproutvideo.com/file/799dd8b41e17e2cdf0/d2fed9bfa257f954/480.mp4"
-          },
-          {"order": 2, "name": "Cadeia abdutora"},
-          {
-            "order": 3,
-            "name": "Eleveção pélvica em colchonete",
-            "link":
-                "https://api-files.sproutvideo.com/file/799dd8b41e17e2cdf0/d2fed9bfa257f954/480.mp4"
-          },
-          {"order": 4, "name": "Glúteo máquina"}
-        ]
-      },
-      {
-        "level": "SOFT",
-        "description":
-            "Carga moderada e movimentos lentos e concentrados para otimizar o padrão de movimento e seus resultados",
-        "frequency": {"repetition": 9, "series": 4},
-        "exercises": [
-          {
-            "order": 1,
-            "name": "Agachamento no smith + agachamento unilateral no simith",
-            "features": [
-              {
-                "name": "BI-SET 4x9 + 9",
-                "description":
-                    "Significa realizar dois movimentos seguidos sem intervalos á cada repetição"
-              }
-            ]
-          },
-          {
-            "order": 2,
-            "name": "Glúteo em 4 apoios com caneleira",
-            "link":
-                "https://api-files.sproutvideo.com/file/799dd8b41e17e2cdf0/d2fed9bfa257f954/480.mp4"
-          },
-          {"order": 3, "name": "Agachamento em polia"},
-          {"order": 4, "name": "Elevação pélvica na máquina 4x até a falha"}
-        ]
-      },
-      {
-        "level": "HARD",
-        "description":
-            "Movimentos lentos e concentrados para otimizar o padrão de movimentos e seus resultados",
-        "frequency": {"repetition": 11, "series": 4},
-        "exercises": [
-          {
-            "order": 1,
-            "name": "Agachamento no smith + agachamento sumô com halteres",
-            "features": [
-              {
-                "name": "BI-SET 4x9 + falha",
-                "description":
-                    "Significa realizar dois movimentos seguidos sem intervalos á cada repetição"
-              }
-            ]
-          },
-          {"order": 2, "name": "Glúteo em apoios com caneleira"},
-          {
-            "order": 3,
-            "name":
-                "Abdução de quadril em pé com caneleira na polia + glúteo na polia",
-            "features": [
-              {
-                "name": "BI-SET 4x9 + 9",
-                "description":
-                    "Significa realizar dois movimentos seguidos sem intervalos á cada repetição"
-              }
-            ]
-          },
-          {"order": 4, "name": "Agachamento búlgaro 4x até a falha"}
-        ]
-      }
-    ]
-  };
-
-  static const bicepsDePedra = {
-    "id": "ret3459i39",
-    "name": "Bíceps de pedra",
-    "description": "Para voçê que busca bíceps tão duro quanto uma pedra.",
-    "cover":
-        "https://firebasestorage.googleapis.com/v0/b/myself-dg.appspot.com/o/powerflix%2Fbiceps_de_pedra.jpg?alt=media&token=d1989712-5e43-4970-8191-d506d165b8b8",
-    "modules": [
-      {
-        "level": "LIGHT",
-        "description":
-            "Movimentos lentos e concentrados para otimizar o padrão de movimento e seus resultados",
-        "frequency": {"repetition": 11, "series": 3},
-        "exercises": [
-          {"order": 1, "name": "Bíceps com barra na polia"},
-          {"order": 2, "name": "Bíceps direto com halteres sentado"},
-          {"order": 3, "name": "Bíceps martelo com halteres"},
-          {"order": 4, "name": "Bíceps scott"}
-        ]
-      },
-      {
-        "level": "SOFT",
-        "description":
-            "Carga moderada e movimentos lentos e concentrados para otimizar o padrão de movimento e seus resultados",
-        "frequency": {"repetition": 9, "series": 4},
-        "exercises": [
-          {
-            "order": 1,
-            "name": "Bíceps com barra na polia + bíceps martelo com halters",
-            "features": [
-              {
-                "name": "BI-SET 4x9 + 9",
-                "description":
-                    "Significa realizar dois movimentos seguidos sem intervalos á cada repetição"
-              }
-            ]
-          },
-          {"order": 2, "name": "Bíceps alternado com halteres sentado"},
-          {"order": 3, "name": "Bíceps concentrado com halteres"},
-        ]
-      },
-      {
-        "level": "HARD",
-        "description":
-            "Movimentos lentos e concentrados para otimizar o padrão de movimentos e seus resultados",
-        "frequency": {"repetition": 6, "series": 4},
-        "exercises": [
-          {"order": 1, "name": "Bíceps martelo com corda na polia"},
-          {"order": 2, "name": "Bíceps alternado com isometria"},
-          {
-            "order": 3,
-            "name": "Bíceps com barra reta rest pause",
-            "features": [
-              {
-                "name": "BI-SET 4x6 + falha",
-                "description":
-                    "Significa realizar dois movimentos seguidos sem intervalos á cada repetição"
-              }
-            ]
-          },
-          {"order": 4, "name": "Bíceps 21 com barra w"}
-        ]
-      }
-    ]
-  };
-
-  static const abdomenChapado = {
-    "id": "egywe8gfs87gf7",
-    "name": "Abdômen Chapado",
-    "description": "Para voçê que busca abdômen top.",
-    "cover":
-        "https://firebasestorage.googleapis.com/v0/b/myself-dg.appspot.com/o/powerflix%2Fabd_chapado.jpg?alt=media&token=b0fb9970-d6d7-4b34-ba92-df7f55059984",
-    "modules": [
-      {
-        "level": "LIGHT",
-        "description":
-            "Movimentos lentos e concentrados para otimizar o padrão de movimento e seus resultados",
-        "frequency": {"repetition": 11, "series": 3},
-        "exercises": [
-          {"order": 1, "name": "Abd supra"},
-          {"order": 2, "name": "Abd infra no chão"},
-          {"order": 3, "name": "Abd supra com pé alto"},
-          {"order": 4, "name": "Abd tesoura em colchonete"}
-        ]
-      },
-      {
-        "level": "SOFT",
-        "description":
-            "Carga moderada e movimentos lentos e concentrados para otimizar o padrão de movimento e seus resultados",
-        "frequency": {"repetition": 9, "series": 4},
-        "exercises": [
-          {"order": 1, "name": "Abd bicicleta em colchonete"},
-          {
-            "order": 2,
-            "name": "Abd escalador em colchonete",
-            "features": [
-              {
-                "name": "3x15",
-                "description":
-                    "Para esse exercício especifico fazer 3 series de 15 repetições."
-              }
-            ]
-          },
-          {"order": 3, "name": "Abd canivete"},
-          {
-            "order": 4,
-            "name": "Abd canoa isométrica",
-            "features": [
-              {
-                "name": "3x15",
-                "description":
-                    "Para esse exercício especifico fazer 3 series de 15 repetições."
-              }
-            ]
-          },
-          {"order": 5, "name": "Abd infra + quadril em colchonete"},
-        ]
-      },
-      {
-        "level": "HARD",
-        "description":
-            "Movimentos lentos e concentrados para otimizar o padrão de movimentos e seus resultados",
-        "frequency": {"repetition": 0, "series": 4},
-        "exercises": [
-          {"order": 1, "name": "Abd prancha lateral em colchonete"},
-          {"order": 2, "name": "Abd prancha ventral isométrica"},
-          {"order": 3, "name": "Abd tapa calcanhar"},
-          {"order": 4, "name": "Abd remador em colchonete"}
-        ]
-      }
-    ]
-  };
-}
