+++ lib/core/domain/model.dart
@@ -1,0 +1,7 @@
+abstract class Model {
+  const Model();
+
+  Map<String, dynamic> toMap();
+
+  String toJson();
+}
