+++ lib/shared/theme/colors.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+const difficultyColors = <Color>[
+  Color(0xFF30D407), // light
+  Color(0xFFEBE109), // soft
+  Color(0xFFEF4040), // hard
+];
